feat(Input.tsx): поля формы и их стили вынесены в отдельный компонент
